back: refactor db connection initialization in all back files
