+++ back/app.js
@@ -3,10 +3,7 @@ const cookieParser = require('cookie-parser')
 const session = require('express-session')
 const config = require('config')
 const Raven = require('raven')
-const { Model } = require('objection')
-const pg = require('pg')
 
-const Knex = require('knex')
 const morgan = require('morgan')
 const helmet = require('helmet')
 const pgConnectSimple = require('connect-pg-simple')
@@ -28,21 +25,7 @@ const declarationMonthsRouter = require('./routes/declarationMonths')
 const employersRouter = require('./routes/employers')
 const developerRouter = require('./routes/developer')
 
-/* https://github.com/tgriesser/knex/issues/927
- * This tells node-pg to use float type for decimal
- * which it does not do because JS loses precision on
- * big decimal number.
- * For our usage (salary), this is not an issue.
- */
-const PG_DECIMAL_OID = 1700
-pg.types.setTypeParser(PG_DECIMAL_OID, parseFloat)
-
-const knex = Knex({
-  client: 'pg',
-  useNullAsDefault: true,
-  connection: process.env.DATABASE_URL,
-})
-Model.knex(knex)
+require('./lib/db') // setup db connection
 
 const isDevEnv = process.env.NODE_ENV === 'development'
 const isProd = process.env.NODE_ENV === 'production'
