--- back/knexfile.js
@@ -1,4 +1,0 @@
-module.exports = {
-  client: 'pg',
-  connection: process.env.DATABASE_URL,
-}
