+++ back/mailing-agent.js
@@ -1,8 +1,7 @@
 const config = require('config')
 const { job } = require('cron')
-const { Model } = require('objection')
-const Knex = require('knex')
-const pg = require('pg')
+
+require('./lib/db') // setup db connection
 
 if (
   !config.get('shouldSendCampaignEmails') &&
@@ -11,22 +10,6 @@ if (
   console.log('Mailing Agent is deactivated.')
   process.exit()
 }
-
-/* https://github.com/tgriesser/knex/issues/927
- * This tells node-pg to use float type for decimal
- * which it does not do because JS loses precision on
- * big decimal number.
- * For our usage (salary), this is not an issue.
- */
-const PG_DECIMAL_OID = 1700
-pg.types.setTypeParser(PG_DECIMAL_OID, parseFloat)
-
-const knex = Knex({
-  client: 'pg',
-  useNullAsDefault: true,
-  connection: process.env.DATABASE_URL,
-})
-Model.knex(knex)
 
 const sendDeclarationCampaign = require('./lib/mailings/sendDeclarationCampaign')
 const sendDeclarationReminderCampaign = require('./lib/mailings/sendDeclarationReminderCampaign')
