--- back/tests/getKnexTestInstance.js
@@ -1,2 +1,0 @@
-const Knex = require('knex')
-module.exports = Knex(require('../knexfile'))
